controllers: remove legacy test suite
As it no longer contains any test cases.
